@@ -2,7 +2,7 @@
 
 	<PropertyGroup>
 		<LangVersion>9.0</LangVersion>
-		<TargetFrameworks>net7.0;net6.0;netcoreapp3.1;net48;netstandard2.1;netstandard2.0</TargetFrameworks>
+		<TargetFrameworks>net8.0;net7.0;net6.0;netcoreapp3.1;net48;netstandard2.1;netstandard2.0</TargetFrameworks>
 		<AssemblyName>ExCSS</AssemblyName>
 		<PackageId>Anateus.ExCSS</PackageId>
 		<Title>ExCSS .NET Stylesheet Parser</Title>
